+++ docs/examples/internationalization/accept_language.py
@@ -1,3 +1,5 @@
+from django.utils.translation import gettext_lazy as _
+
 from dmr import Controller, ResponseSpec
 from dmr.exceptions import NotAuthenticatedError
 from dmr.plugins.pydantic import PydanticSerializer
@@ -8,6 +10,7 @@ class UsersController(Controller[PydanticSerializer]):
         ResponseSpec(
             Controller.error_model,
             status_code=NotAuthenticatedError.status_code,
+            description=_('Not authenticated'),  # translated via django way
         ),
     )
 
@@ -17,3 +20,4 @@ class UsersController(Controller[PydanticSerializer]):
 
 
 # run: {"controller": "UsersController", "method": "post", "headers": {"Accept-Language": "ru"}, "url": "/api/lang/", "assert-error-text": "security", "fail-with-body": false}  # noqa: ERA001, E501
+# openapi: {"controller": "UsersController", "openapi_url": "/docs/openapi.json"}  # noqa: ERA001, E501